Add files via upload
This unity tutorial had many helpful tips and tricks, and was useful in figuring out things I wanted to avoid.
During this assignment the guide used a canvas to set a background, which caused alot of problems
These problems included the resizing of the program, and this was extremely a problem with the collider boxes
After sometime I was able to overcome the collider box issue, however the resizing of the cards and background
was to much of a hassle, and I'll be using gameobjects moving forward. Also trying to reuse scripts was a concept I didn't
consider while writing code. I think the idea of having cards work on a script and layout all the cards once they are drawn, isn’t going to work for blackjack. 
For the next assignment I’ll more than likely keep the card decks in a do-while script. 
This will allow me to end the loop once all the cards have been drawn. The problem will be between the “enemy” dealing cards, and them interacting with the player cards. 
Maybe using temp variables always on the player, and having the cards pass through the player like the cards will be passing through a function. 
So like the player is given cards in by the dealer, and is than reset at the start of every round.